Moved classes to separate files.
