+++ BambooHR/API/BambooHTTP.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * BambooHTTP.php
+ */
+
+namespace BambooHR\API;
+
+/**
+ * BambooHTTP
+ * @package BambooHR
+ */
+interface BambooHTTP {
+    function setBasicAuth($username, $password);
+    function sendRequest(BambooHTTPRequest $request);
+}
